Added delt_max to the namelist.
Before the maximum time step was the <delt> given in the <knob> namelist. However when we restart we do not want the restarted time step to be the maximum time step. Therefore I added a new variable <delt_max> which will set the maximum time step. If it is not mentioned, the previous before of code_delt_max = delt will be retained.