Undid my last edit
I was looking at the code from 13 December and noticed that get_fluxes() doesn't take h as an input, but rather g. This routine had been taking g as an input since the edit on 28 Feb 2018, which is why I removed the comment. However, I noticed Micheal changed g to h again on his edit of 15 December 2021, so I'm putting the comment back. I do not understand though why it has been changed back to h, we should be able to calculate the fluxes based on g, so why would we add extra operations to switch g to h and back, if it is not needed? Also, if h is passed onto the routine, why not rewrite it to subroutine get_fluxes(h), to avoid confusion?