Fix bug with restarted simulations with longer input files
Fixes #85

Previously, if a simulation was restarted and the new input file was
longer than the one in the netCDF file, we would hard error.
Now, the "number of lines" dimension is unlimited, so longer input
files can always be written. If the new input file is shorter, we
blank out the variable first.