Issue with linear simulations: first ky in [aky_min,...,aky_max] has phi=omega=gamma=0
When running a linear simulation, the output for the first ky in a range of naky selected ky's ([aky_min,...,aky_max]) will be zero, for gamma, for omega, for phi, ... This is due to the "noise" initialisation which sets phi(1,1,:,:)=0 and since the simulations are linear the potential remains zero throughout the simulation. In order to fix this issue, I force the "default" initialisation to be used when a linear simulation is run.