Add gamtot3 to FFS
Adiabatic response matrix requires gamtot3. Have added in a 'hack' for now - needs to be changed later. At the moment this is just using the first alpha value for dl_over_b when calculating gamtot3, but this needs to be changed to an appropriate average. Has been added for now to allow clean up to move forward.