Merge Full Flux Surface code into Master with Electromagnetic effects (#135)
* Merge in Full Flux Surface code

* Add gamtot3 to FFS

Adiabatic response matrix requires gamtot3. Have added in a 'hack' for now - needs to be changed later. At the moment this is just using the first alpha value for dl_over_b when calculating gamtot3, but this needs to be changed to an appropriate average. Has been added for now to allow clean up to move forward.

* Add correction term

Changed the centering for correction terms to be more consistent with implicit_solve.f90. Also added in missing correction term for FFS

* Phase factor

Add phase factor to response matrix when periodic (potentially needed for FFS)

* Fix adiabatic FFS

* Initialise all arrays in miller_geometry.f90 to zero

* Initialise all geometry arrays to zero in stella_geometry.f90

* If not FFS, set rhostar to 0.

* Add mp_abort for FFS with kinetic electrons

---------

Co-authored-by: HanneThienpondt <Hanne.Thienpondt@outlook.com>